+++ server/lib/yacht.h
@@ -1,3 +1,3 @@
-double *generate_rotation(int num, int *r_result, double *r_translations,
-                         double *r_rotations);
-void free_ptr(double *ptr);
+float *generate_rotation(int num, int *r_result, float *r_translations,
+                         float *r_rotations);
+void free_ptr(float *ptr);
